관심 종목 화면 - 가격/등락/거래량 표시

+++ app/src/main/java/com/trueedu/spac/ui/following/views/FollowingItem.kt
@@ -37,7 +37,7 @@ fun FollowingItem(
     low: Double?,
     delta: Double?,
     rate: Double?,
-    volume: Double,
+    volume: Long,
     halt: Boolean,
     designated: Boolean,
     hasDisclosure: Boolean,
